moved everything to dynamic imports for better performance
